holiday upsolving

+++ Algo/Graphs/dijkstra.cpp
@@ -1,12 +1,12 @@
 //O((V+E) logE) tested https://cses.fi/problemset/task/1671
 const int N = 2e5+3;
-vector<int> adj[N], adjw[N];
-int dist[N];
+vector<ll> adj[N], adjw[N];
+ll dist[N];
 
 void dijkstra(int src){
     memset(dist, 63, sizeof(dist));
     dist[src] = 0;
-    priority_queue<pair<int,int>, vector<pair<int,int>>, greater<pair<int,int>>> pq;
+    priority_queue<pair<ll,ll>, vector<pair<ll,ll>>, greater<pair<ll,ll>>> pq;
     pq.push({0, src});
     while(!pq.empty()){
         auto d = pq.top().first;
@@ -14,7 +14,7 @@ void dijkstra(int src){
         pq.pop();
         if(d > dist[u]) continue;
         for(int i=0;i<adj[u].size();i++){
-            int v = adj[u][i], w = adjw[u][i];
+            ll v = adj[u][i], w = adjw[u][i];
             if(dist[v] > dist[u] + w){
                 dist[v] = dist[u] + w;
                 pq.push({dist[v], v});
